Start of refactoring interface, state, and data. Interface now updates state independantly

+++ SubZero/test/VideoTesting.cpp
@@ -104,18 +104,18 @@ cv::Mat Moments(cv::Mat img){
 
     /// Get the moments
     std::vector<cv::Moments> mu(contours.size() );
-    for( int i = 0; i < contours.size(); i++ )
+    for( uint64_t i = 0; i < contours.size(); i++ )
     { mu[i] = moments( contours[i], false ); }
 
     ///  Get the mass centers:
     std::vector<cv::Point2f> mc( contours.size() );
-    for( int i = 0; i < contours.size(); i++ )
+    for( uint64_t i = 0; i < contours.size(); i++ )
     { mc[i] = cv::Point2f( mu[i].m10/mu[i].m00 , mu[i].m01/mu[i].m00 ); }
 
 
     /// Draw contours
     cv::Mat drawing = cv::Mat::zeros( canny_output.size(), CV_8UC3 );
-    for( int i = 0; i< contours.size(); i++ )
+    for( uint64_t i = 0; i< contours.size(); i++ )
     {
         cv::Scalar color = cv::Scalar(0, 255, 0);
         drawContours( drawing, contours, i, color, 2, 8, hierarchy, 0, cv::Point() );
@@ -125,7 +125,7 @@ cv::Mat Moments(cv::Mat img){
 
     /// Calculate the area with the moments 00 and compare with the result of the OpenCV function
     //printf("\t Info: Area and Contour Length \n");
-    for( int i = 0; i< contours.size(); i++ )
+    for( uint64_t i = 0; i< contours.size(); i++ )
     {
         //printf(" * Contour[%d] - Area (M_00) = %.2f - Area OpenCV: %.2f - Length: %.2f \n", i, mu[i].m00, contourArea(contours[i]), arcLength( contours[i], true ) );
         cv::Scalar color = cv::Scalar(255, 0, 0);
